Move FINISH descr classes from majit-metainterp::compile to majit-backend::finish_descrs
compile.py:618-672, 1092-1099 _DoneWithThisFrameDescr family,
ExitFrameWithExceptionDescrRef, and PropagateExceptionDescr depend only
on majit-ir traits and std primitives; nothing in their definition
requires majit-metainterp.  Hosting them inside majit-metainterp
prevented backend codegen and synthetic-exit paths from constructing
the proper class-distinct types directly.

Move the six classes (and the shared DoneWithThisFrameDescrBase) to a
new majit-backend::finish_descrs module.  majit-metainterp re-exports
them from compile.rs so existing crate::compile::DoneWithThisFrameDescr*
references keep resolving.
